requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.20
httpx>=0.27.0
pgvector>=0.3.0
reductoai>=0.1.0
pymupdf>=1.24.0
python-docx>=1.0.0
